Move trackPaintGeneric and helper functions to new files
