+++ src/openrct2/libopenrct2.vcxproj
@@ -418,6 +418,7 @@
     <ClInclude Include="paint\track\coaster\WoodenRollerCoaster.hpp" />
     <ClInclude Include="paint\track\Segment.h" />
     <ClInclude Include="paint\track\Support.h" />
+    <ClInclude Include="paint\track\TrackPaintGeneric.h" />
     <ClInclude Include="paint\track_pieces\QuarterHelix.h" />
     <ClInclude Include="paint\vehicle\Vehicle.MiniGolf.h" />
     <ClInclude Include="paint\vehicle\VehiclePaint.h" />
@@ -1086,6 +1087,7 @@
     <ClCompile Include="paint\track\water\RiverRapids.cpp" />
     <ClCompile Include="paint\track\water\SplashBoats.cpp" />
     <ClCompile Include="paint\track\water\SubmarineRide.cpp" />
+    <ClCompile Include="paint\track\TrackPaintGeneric.cpp" />
     <ClCompile Include="paint\track_pieces\QuarterHelix.cpp" />
     <ClCompile Include="paint\vehicle\VehiclePaint.cpp" />
     <ClCompile Include="paint\vehicle\Vehicle.LaunchedFreefall.cpp" />
